+++ jimVermillionFE/app/css/layout.css
@@ -1,0 +1,41 @@
+main {
+  overflow: hidden;
+}
+header, section {
+  margin: 10px;
+  border: 1px solid red;
+}
+footer {
+  height: 50px;
+  width: 100%;
+  text-align: center;
+  padding: 10px;
+}
+header h1 {
+  text-align: center;
+  font-size: 3em;
+  color: #4C4B63;
+}
+.right {
+  float: right;
+  width: 30%;
+  text-align: center;
+}
+.left {
+  float: left;
+  width: 30%;
+  text-align: center;
+}
+input, label {
+  display: block;
+  margin: 10px auto;
+}
+.nonCrud {
+  text-align: center;
+  margin: 0 auto;
+  width: 30%;
+  font-size: 36px;
+}
+button {
+  margin-bottom: 10px;
+}
